delivery_price тип изменен на decimal для запроса на обновление привязки к городу

+++ flora_api_client/presentations/partners.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from dataclasses import dataclass, field
 
 from marshmallow.validate import Range, Length
@@ -22,7 +23,7 @@ class Partner(User):
 @dataclass
 class BindCityRequestDataclass(BaseDataclass):
     city_id: int = field(metadata={"strict": True})
-    delivery_price: int = field(metadata={"validate": Range(min=0)})
+    delivery_price: Decimal = field(metadata={"validate": Range(min=0)})
     delivery_currency: str = field(metadata={"validate": Length(equal=3)})
